feat: refactor playground
