+++ learnings/5/index.ts
