+++ package.json
@@ -17,7 +17,7 @@
     "start": "parcel src/index.html",
     "test": "yarn test:build && karma start karma.conf.cjs",
     "test:watch": "nodemon --exec \"yarn test\" --watch src --watch tests -e js,ts",
-    "test:build": "rimraf dist-test/* && parcel build '(src|tests)/**/*.test.ts' --dist-dir dist-test",
+    "test:build": "rimraf dist-test/* && parcel build '(src|tests)/**/*.test.ts' --dist-dir dist-test --no-optimize",
     "build": "yarn generate:version && parcel build src/index.html --public-url .",
     "deploy": "yarn build && gh-pages -d ./dist",
     "generate:version": "genversion src/version.gen.ts --es6",
@@ -34,6 +34,7 @@
     "karma-spec-reporter": "^0.0.33",
     "nodemon": "^2.0.15",
     "parcel": "^2.1.1",
+    "process": "^0.11.10",
     "puppeteer": "^13.4.0",
     "rimraf": "^3.0.2",
     "ts-node": "^10.4.0",
