+++ src/index.html
@@ -7,11 +7,14 @@
     <title>CodeMirror6 - Block Indentation</title>
   </head>
   <body>
+    <div>
+      <h1 style="display: inline">Block-based Text Editor</h1>
+      <span style="float: right">v<span id="version"></span></span>
+    </div>
     <div id="editor"></div>
     <hr />
     <h3>Events</h3>
     <div id="notifications"></div>
-    <div style="float: right">v<span id="version"></span></div>
     <script src="./index.ts" type="module"></script>
   </body>
 </html>
